Add importance level to block lock screen notifications (fixed) (2/2)
Adjust for XOS 7.1

One feature that is still missing in the handling of notifications in
N is the possibility to block them app-wise for relevance/importance
reasons rather than for privacy reasons – i.e. also on the insecure
lock screen.

This patch plugs the hole by introducing a new importance level "very
low" between "min" and "low" in the power notification controls that
allows notifications on the status bar, but not on the lock screen.

It brings back most of the functionality of
Ib166db1b1673aeaea132c8eeb16c650d2f254a82 excluding the ability of
blocking only ongoing (persistent) notifications.

This fixed version does not break the API.

Change-Id: I394d88ef516025998b3e31a191feeb5b08b85446
Signed-off-by: Marc K <morckx@gmail.com>
Signed-off-by: Harsh Shandilya <msfjarvis@gmail.com>
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>