DeviceInfoSettings : More robust handling in setting revision
To set the summary for KEY_REVISIONS, SystemProperties#get(String)
is used to get the value of ro.mod.revisions, which is then split
on a space delimiter and parsed into separate AOSP and CAF revisions.

Before this patch, we were assuming utopia that the format of
ro.mod.revisions matches what we expect, which was not the case
when I pushed ONLY the Settings module onto a system with the older
prop format, raising an ArrayOutOfBoundsException every time I
tried to launch this Fragment. Fix that by catching the exception
and removing the PreferenceScreen itself, thus allowing other functions
to be normally used.

Change-Id: I19d726d1376f3cc410096b0022051ae4fbd7c690