fix contacts access toggle in pairing dialog for BT hands-free devices
There was a mismatch between checks in getContactSharingState() and setContactSharingState(),
which led to contacts access being granted by default when the user accepted a background
pairing request for a hands-free device, despite the contacts access toggle indicating that the
contacts access will not be granted.

Fix this check mismatch by removing the confusing setContactSharingState() method and switching code
that depends on setContactSharingState() to use the value of getContactSharingState() instead.

Signed-off-by: Simão Gomes Viana <devel@superboring.dev>